backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
anthropic==0.42.0
pymupdf==1.25.3
python-pptx==1.0.2
python-dotenv==1.0.1
aiofiles==24.1.0
httpx==0.28.1
bcrypt==4.2.1
pyjwt==2.10.1
